requirements.txt
# Core Framework
flask==2.3.3
flask-restful==0.3.10
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Database
flask-sqlalchemy==3.1.1
alembic==1.12.0
pymongo==4.5.0

# Authentication
flask-jwt-extended==4.5.3
passlib==1.7.4

# AI and NLP
openai==1.5.0
anthropic==0.5.0
langchain==0.0.310
transformers==4.34.0
sentence-transformers==2.2.2
chromadb==0.4.15

# Publishing Integrations
linkedin-api==2.0.0a3
medium-api==0.4.1
tweepy==4.14.0

# Utilities
pydantic==2.4.2
tenacity==8.2.3
pytest==7.4.2
requests==2.31.0
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.0